move gaussian filter definition as in skimage 0.19
