+++ python/cucim/src/cucim/skimage/feature/_canny.py
@@ -18,8 +18,8 @@ import cupyx.scipy.ndimage as ndi
 from cupyx.scipy.ndimage import binary_erosion, generate_binary_structure
 
 from .. import dtype_limits, img_as_float
+from .._shared.filters import gaussian
 from .._shared.utils import check_nD
-from ..filters import gaussian
 
 
 # fuse several commonly paired ufunc operations into a single kernel call
