+++ python/cucim/src/cucim/skimage/feature/_daisy.py
@@ -1,9 +1,9 @@
 import math
 
 import cupy as cp
-from cupyx.scipy.ndimage import gaussian_filter
 
 from .. import img_as_float
+from .._shared.filters import gaussian
 from .._shared.utils import check_nD
 
 
@@ -144,9 +144,8 @@ def daisy(image, step=4, radius=15, rings=3, histograms=8, orientations=8,
     hist_smooth = cp.empty((rings + 1,) + hist.shape, dtype=float_dtype)
     for i in range(rings + 1):
         for j in range(orientations):
-            hist_smooth[i, j, :, :] = gaussian_filter(hist[j, :, :],
-                                                      sigma=sigmas[i],
-                                                      mode='reflect')
+            hist_smooth[i, j, :, :] = gaussian(hist[j, :, :], sigma=sigmas[i],
+                                               mode='reflect')
 
     # Assemble descriptor grid.
     theta = [2 * pi * j / histograms for j in range(histograms)]
